// frontend/src/app/modules/dips/types/order.ts
export interface Order {
  field?: string;
  direction?: string;
}

export enum OrderField {
  Number = 'dip',
  Title = 'title',
  Summary = 'sentenceSummary',
  Status = 'status',
  MostUsed = 'subproposalsCount',
}

export enum OrderDirection {
  ASC = '',
  DESC = '-',
}

export enum OrderFieldName {
  Number = 'Number',
  Title = 'Title',
  Summary = 'Summary',
  Status = 'Status',
  MostUsed = 'MostUsed',
}